pyproject.toml
[project]
name = "automatisierung"
version = "0.0.0-dev"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "astral~=3.2",
    "caldav~=2.0.0",
    "icalevents~=0.2.1",
    "imap-tools~=1.10.0",
    "mgrs~=1.5.0",
    "paho-mqtt~=2.1.0",
    "pycryptodome~=3.22.0",
    "python-socketio~=5.13.0",
    "requests~=2.32.4",
    "schedule~=1.2.2",
    "toml~=0.10.2",
    "typeguard~=4.4.4",
    "watchdog~=6.0.0",
]
